Use cupti's internal crc for cubins

+++ src/tool/hpcrun/gpu/nvidia/cubin-id-map.c
@@ -76,7 +76,7 @@
 //*****************************************************************************
 
 struct cubin_id_map_entry_s {
-  uint32_t cubin_id;
+  uint64_t cubin_id;
   uint32_t hpctoolkit_module_id;
   Elf_SymbolVector *elf_vector;
   struct cubin_id_map_entry_s *left;
@@ -85,7 +85,7 @@ struct cubin_id_map_entry_s {
 
 
 typedef struct {
-  uint32_t cubin_id;
+  uint64_t cubin_id;
   cubin_id_map_entry_t *entry;
 } cubin_id_map_hash_entry_t;
 
@@ -105,7 +105,7 @@ static spinlock_t cubin_id_map_lock = SPINLOCK_UNLOCKED;
 //*****************************************************************************
 
 static cubin_id_map_entry_t *
-cubin_id_map_entry_new(uint32_t cubin_id, Elf_SymbolVector *vector)
+cubin_id_map_entry_new(uint64_t cubin_id, Elf_SymbolVector *vector)
 {
   cubin_id_map_entry_t *e;
   e = (cubin_id_map_entry_t *)hpcrun_malloc_safe(sizeof(cubin_id_map_entry_t));
@@ -119,7 +119,7 @@ cubin_id_map_entry_new(uint32_t cubin_id, Elf_SymbolVector *vector)
 
 
 static cubin_id_map_entry_t *
-cubin_id_map_splay(cubin_id_map_entry_t *root, uint32_t key)
+cubin_id_map_splay(cubin_id_map_entry_t *root, uint64_t key)
 {
   REGULAR_SPLAY_TREE(cubin_id_map_entry_s, root, key, cubin_id, left, right);
   return root;
@@ -151,7 +151,7 @@ cubin_id_map_delete_root()
 cubin_id_map_entry_t *
 cubin_id_map_lookup
 (
- uint32_t id
+ uint64_t id
 )
 {
   cubin_id_map_entry_t *result = NULL;
@@ -191,7 +191,7 @@ cubin_id_map_lookup
 void
 cubin_id_map_insert
 (
- uint32_t cubin_id,
+ uint64_t cubin_id,
  uint32_t hpctoolkit_module_id,
  Elf_SymbolVector *vector
 )
@@ -236,7 +236,7 @@ cubin_id_map_insert
 void
 cubin_id_map_delete
 (
- uint32_t cubin_id
+ uint64_t cubin_id
 )
 {
   cubin_id_map_root = cubin_id_map_splay(cubin_id_map_root, cubin_id);
@@ -271,11 +271,11 @@ cubin_id_map_entry_elf_vector_get
 // looking up elf symbols inside a cubin
 //--------------------------------------------------------------------------
 ip_normalized_t
-cubin_id_transform(uint32_t cubin_id, uint32_t function_index, uint64_t offset)
+cubin_id_transform(uint64_t cubin_id, uint32_t function_index, uint64_t offset)
 {
   cubin_id_map_entry_t *entry = cubin_id_map_lookup(cubin_id);
   ip_normalized_t ip;
-  TMSG(CUDA_CUBIN, "cubin_id %d", cubin_id);
+  TMSG(CUDA_CUBIN, "cubin_id %lu", cubin_id);
   if (entry != NULL) {
     uint32_t hpctoolkit_module_id = cubin_id_map_entry_hpctoolkit_id_get(entry);
     TMSG(CUDA_CUBIN, "get hpctoolkit_module_id %d", hpctoolkit_module_id);
