+++ src/tool/hpcrun/gpu/nvidia/cubin-id-map.h
@@ -76,14 +76,14 @@ typedef struct cubin_id_map_entry_s cubin_id_map_entry_t;
 cubin_id_map_entry_t *
 cubin_id_map_lookup
 (
- uint32_t cubin_id
+ uint64_t cubin_id
 ); 
 
 
 void
 cubin_id_map_insert
 (
- uint32_t cubin_id, 
+ uint64_t cubin_id, 
  uint32_t hpctoolkit_module_id, 
  Elf_SymbolVector *vector
 );
@@ -92,7 +92,7 @@ cubin_id_map_insert
 void
 cubin_id_map_delete
 (
- uint32_t cubin_id
+ uint64_t cubin_id
 );
 
 
@@ -113,7 +113,7 @@ cubin_id_map_entry_elf_vector_get
 ip_normalized_t
 cubin_id_transform
 (
- uint32_t cubin_id, 
+ uint64_t cubin_id, 
  uint32_t function_id, 
  uint64_t offset
 );
